requirements.txt
langchain>=0.3.0
langchain-openai>=0.2.0
langchain-anthropic>=0.1.0
langchain-community>=0.3.0
langgraph>=0.2.0
openai>=1.40.0
chromadb>=0.5.0
PyMuPDF>=1.24.0
pydantic>=2.7.0
SQLAlchemy>=2.0.0
psycopg2-binary>=2.9.0
typing-extensions>=4.9.0
python-dotenv>=1.0.0
pandas>=2.0.0
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
apache-airflow>=2.7.0
reportlab>=4.0.0
pytest>=8.0.0
